+++ lib/ventriloquist/cap/linux/rvm_install.rb
@@ -4,10 +4,10 @@ module VagrantPlugins
       module Linux
         module RvmInstall
           def self.rvm_install(machine)
-            if ! machine.communicate.test('test -d $HOME/.rvm')
+            if ! machine.communicate.test('test -d /usr/local/rvm')
               machine.env.ui.info('Installing RVM')
-              machine.communicate.execute('\curl -L https://get.rvm.io | bash -s stable --autolibs=enabled --ignore-dotfiles')
-              machine.communicate.execute("echo 'source $HOME/.rvm/scripts/rvm' >> ~/.profile")
+              machine.communicate.sudo('\curl -L https://get.rvm.io | bash -s stable --autolibs=enabled')
+              machine.communicate.sudo("usermod -a -G rvm #{machine.config.ssh.username}")
             end
           end
         end