+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [0.3.3](https://github.com/fgrehm/ventriloquist/compare/v0.3.2...master) (unreleased)
 
+FEATURES:
+
+  - Install RVM globally [[GH-24]]
 
 ## [0.3.2](https://github.com/fgrehm/ventriloquist/compare/v0.3.1...v0.3.2) (October 28, 2013)
 
@@ -51,6 +54,7 @@ BUG FIXES:
 
 
 
+[GH-24]: https://github.com/fgrehm/ventriloquist/issues/24
 [GH-9]: https://github.com/fgrehm/ventriloquist/issues/9
 [GH-19]: https://github.com/fgrehm/ventriloquist/issues/19
 [GH-3]: https://github.com/fgrehm/ventriloquist/issues/3